requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-restx==0.2.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
oauth2==1.9.0.post1
pyrsistent==0.16.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
uWSGI==2.0.19.1
uwsgitop==0.11
Werkzeug==1.0.1
